Moved wrongfully stuff to right location
