--- const.py
@@ -1,3 +1,0 @@
-DOMAIN = "fht_heating"
-DEV_NAME = "dev_name"
-ADDRESS = "address"
