+++ fht_heating/const.py
@@ -1,2 +1,3 @@
+DOMAIN = "fht_heating"
 DEV_NAME = "dev_name"
 ADDRESS = "address"
