+++ fht_heating/manifest.json
@@ -7,5 +7,6 @@
     "requirements": ["beautifulsoup4", "lxml", "requests"],
     "iot_class": "local_polling",
     "version": "0.2.0",
-    "homeassistant": "2025.1.0"
+    "homeassistant": "2025.1.0",
+    "config_flow": true
 }
