--- manifest.json
@@ -1,12 +1,0 @@
-{
-    "domain": "fht_heating",
-    "name": "FHT Heating",
-    "documentation": "",
-    "dependencies": [],
-    "codeowners": [],
-    "requirements": ["beautifulsoup4", "lxml", "requests"],
-    "iot_class": "local_polling",
-    "version": "0.2.0",
-    "homeassistant": "2025.1.0",
-    "config_flow": true
-}
