--- translations/en.json
@@ -1,3 +1,0 @@
-{
-  "title": "FHT Heating"
-}
